Clients Can Connect
The CCC Update
+Clients can connect and move. With camera difficulty
+Clients need to sync position with each other
