+++ SMNC/Assets/Prefabs/Character/Character.prefab
@@ -205,7 +205,6 @@ GameObject:
   - component: {fileID: 1341460666444318548}
   - component: {fileID: 1341460666444318551}
   - component: {fileID: -8425304616194810600}
-  - component: {fileID: -9068046612596395886}
   m_Layer: 0
   m_Name: Character
   m_TagString: Untagged
@@ -224,7 +223,6 @@ Transform:
   m_LocalPosition: {x: 5.188416, y: 5.999, z: 4.747}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
-  - {fileID: 7676315692826096792}
   - {fileID: 1341460666444487355}
   - {fileID: 1341460666444487357}
   m_Father: {fileID: 0}
@@ -278,6 +276,8 @@ MonoBehaviour:
   gravity: -9.8
   jumpHeight: 5
   floatingJumpModifier: 1.5
+  horizontal: 0
+  vertical: 0
 --- !u!114 &1341460666444318550
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -285,13 +285,13 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1341460666444318553}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: ec3b42e173a696f4a8c2d854d202e86b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   sensitivity: 0.5
-  mainCamera: {fileID: 7676315692826096793}
+  mainCamera: {fileID: 0}
   mouseX: 0
   mouseY: 0
 --- !u!114 &1341460666444318548
@@ -301,14 +301,14 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1341460666444318553}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: df103242f1839294c8cd907dcf55846b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   projectile: {fileID: 2010415504639872752, guid: 204f350af7fd5904e838569061554b45, type: 3}
   distanceFromPlayer: 0
-  mainCamera: {fileID: 7676315692826096793}
+  mainCamera: {fileID: 0}
 --- !u!143 &1341460666444318551
 CharacterController:
   m_ObjectHideFlags: 0
@@ -343,20 +343,6 @@ MonoBehaviour:
   AlwaysReplicateAsRoot: 0
   DontDestroyWithOwner: 0
   AutoObjectParentSync: 1
---- !u!114 &-9068046612596395886
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1341460666444318553}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 25d54d2aea43f2647a515654adcb34ba, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  position:
-    m_InternalValue: {x: 0, y: 0, z: 0}
 --- !u!1 &1341460666444318555
 GameObject:
   m_ObjectHideFlags: 0
@@ -386,7 +372,7 @@ Transform:
   m_Children:
   - {fileID: 1341460666444487299}
   m_Father: {fileID: 1341460666444487353}
-  m_RootOrder: 1
+  m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1341460666444318557
 GameObject:
@@ -417,7 +403,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 1341460666444487353}
-  m_RootOrder: 2
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!137 &1341460666439044065
 SkinnedMeshRenderer:
@@ -523,89 +509,6 @@ SkinnedMeshRenderer:
     m_Center: {x: 0.0016462803, y: 0.00000020861626, z: 0.03639815}
     m_Extent: {x: 0.9924226, y: 0.9069464, z: 0.24306375}
   m_DirtyAABB: 0
---- !u!1 &7676315692826096795
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 7676315692826096792}
-  - component: {fileID: 7676315692826096793}
-  - component: {fileID: 7676315692826096794}
-  m_Layer: 0
-  m_Name: Main Camera
-  m_TagString: MainCamera
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &7676315692826096792
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7676315692826096795}
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0.01, y: 1.872, z: -0.31}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 1341460666444487353}
-  m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!20 &7676315692826096793
-Camera:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7676315692826096795}
-  m_Enabled: 1
-  serializedVersion: 2
-  m_ClearFlags: 1
-  m_BackGroundColor: {r: 0.19215687, g: 0.3019608, b: 0.4745098, a: 0}
-  m_projectionMatrixMode: 1
-  m_GateFitMode: 2
-  m_FOVAxisMode: 0
-  m_SensorSize: {x: 36, y: 24}
-  m_LensShift: {x: 0, y: 0}
-  m_FocalLength: 50
-  m_NormalizedViewPortRect:
-    serializedVersion: 2
-    x: 0
-    y: 0
-    width: 1
-    height: 1
-  near clip plane: 0.3
-  far clip plane: 1000
-  field of view: 60
-  orthographic: 0
-  orthographic size: 5
-  m_Depth: -1
-  m_CullingMask:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_RenderingPath: -1
-  m_TargetTexture: {fileID: 0}
-  m_TargetDisplay: 0
-  m_TargetEye: 3
-  m_HDR: 1
-  m_AllowMSAA: 1
-  m_AllowDynamicResolution: 0
-  m_ForceIntoRT: 0
-  m_OcclusionCulling: 1
-  m_StereoConvergence: 10
-  m_StereoSeparation: 0.022
---- !u!81 &7676315692826096794
-AudioListener:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7676315692826096795}
-  m_Enabled: 1
 --- !u!1 &8043197855946136945
 GameObject:
   m_ObjectHideFlags: 0
