+++ SMNC/Assets/Scripts/Player/Movement.cs
@@ -1,56 +1,69 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using Unity.Netcode;
 
-namespace Player
+public class Movement : NetworkBehaviour
 {
-    public class Movement : MonoBehaviour
+    private NetworkVariable<Vector3> networkPosition = new NetworkVariable<Vector3>();
+    [SerializeField] private float moveSpeed;
+
+    // This may need to become/already is a global variable. REMEMBER GRAIVITY IS NEGATIVE PEOPLE. WE. DONT. FLY.
+    [SerializeField] private float gravity; 
+
+    // The force of the jump. (Should represent Unity units of height.)
+    [SerializeField] private float jumpHeight; 
+
+    // Duration of the jump, shorter means the player reaches max height faster.
+
+    // Used for gravity and jumping.
+    private Vector3 playerVerticalVelocity;   
+
+    //floatingJumpModifier is the amount of velocity the player gains by letting go of space during their jump.
+    [SerializeField] private float floatingJumpModifier;  
+    private CharacterController controller;
+    private bool jump; // Whether the jump key has been pressed.
+
+    private Vector3 move;
+
+
+    //Hard falling is toggle on whether or not the player let go of space during their jump.
+    private bool hardFalling;
+    public float horizontal, vertical;
+    private Player player;
+
+
+    public override void OnNetworkSpawn()
     {
-        [SerializeField] private float moveSpeed;
-
-        // This may need to become/already is a global variable. REMEMBER GRAIVITY IS NEGATIVE PEOPLE. WE. DONT. FLY.
-        [SerializeField] private float gravity; 
-
-        // The force of the jump. (Should represent Unity units of height.)
-        [SerializeField] private float jumpHeight; 
-
-        // Duration of the jump, shorter means the player reaches max height faster.
-
-        // Used for gravity and jumping.
-        private Vector3 playerVerticalVelocity;   
-
-        //floatingJumpModifier is the amount of velocity the player gains by letting go of space during their jump.
-        [SerializeField] private float floatingJumpModifier;  
-        private CharacterController controller;
-        private bool jump; // Whether the jump key has been pressed.
-
-        private Vector3 move;
-
-
-        //Hard falling is toggle on whether or not the player let go of space during their jump.
-        private bool hardFalling;
-        public float horizontal, vertical;
-        private Player player;
-
-        // Start is called before the first frame update
-        void Start()
+        if (IsServer)
+            networkPosition.Value = new Vector3(5f, 6f, 5f);
+    }
+    private void Awake()
+    {
+        controller = GetComponent<CharacterController>();
+    }
+    
+    void Start()
+    {
+        if (IsClient && IsOwner)
         {
-            controller = GetComponent<CharacterController>();
-            player = GetComponent<Player>();
             move = Vector3.zero;
             jump = false;
             hardFalling = true;
             playerVerticalVelocity.y = gravity * Time.deltaTime;
         }
+    }
 
-        // Update is called once per frame
-        void Update()
+    
+    void Update()
+    {
+        if (!IsLocalPlayer)
+            return;
+        if (IsClient && IsOwner)
         {
             // Move the character relative to the direction they are facing.
             move = Input.GetAxisRaw("Horizontal") * transform.right 
                 + Input.GetAxisRaw("Vertical") * transform.forward;
-            horizontal = Input.GetAxisRaw("Horizontal");
-            vertical = Input.GetAxisRaw("Vertical");
 
             // If space is pressed, jump.
             if (Input.GetButtonDown("Jump") && controller.isGrounded)
@@ -61,58 +74,71 @@ namespace Player
 
             if (Input.GetButtonUp("Jump"))
                 hardFalling = true;
-        }
-
-        /*
-        * FixedUpdate is where movement/physics based operations should take place.
-        * Do not try to do them in Update().
-        *
-        * move = move.normalized will normalize the movement vector.
-        * This means that the character will move the same speed 
-        * diagonally as the do straight. Because pythagoras.
-        *
-        */
-        void FixedUpdate()
-        {
-            /* 
-            * Jumping is handled by applying a constant velocity to the player for
-            * a duration of time. This is to prevent the player from snapping to
-            * their max jump height, and makes it a bit more smooth. 
-            *
-            * I THINK I have it so jump height represents the maximum number of Unity 
-            * units the player will travel upwards, but don't quote this on that.
-            */
-            
-
-            //if the player is jumping (duh)
-            if (jump)
-            {
-                jump = false;
-                playerVerticalVelocity.y = jumpHeight;
-            }
-            else if (!controller.isGrounded)
-            {
-                //In this else if is when the player is NOT jumping but is still in the air
-                if (hardFalling)
-                    playerVerticalVelocity.y += (gravity * floatingJumpModifier) * Time.deltaTime;
-                else if (Input.GetButton("Jump"))
-                    playerVerticalVelocity.y += gravity * Time.deltaTime;    
-            }
-            else 
-            {
-                //This else is when the player is not jumping and is grounded.
-                playerVerticalVelocity.y = gravity * Time.deltaTime;
-            }
-            Debug.Log(playerVerticalVelocity.y);
-
-            //move = move.normalized;
-            //controller.Move(move * moveSpeed * Time.deltaTime);
-            player.Move(move);
-
-
-            // Add gravity to player's vertical velocity. 
-            //playerVerticalVelocity.y -= gravity;
-            controller.Move(playerVerticalVelocity * Time.deltaTime); 
+            RequestMovementServerRpc(move);
+            transform.position = networkPosition.Value;
         }
     }
+
+    [ServerRpc]
+    public void RequestMovementServerRpc(Vector3 speed)
+    {
+        networkPosition.Value += speed;
+    }
+
+    /*
+    * FixedUpdate is where movement/physics based operations should take place.
+    * Do not try to do them in Update().
+    *
+    * move = move.normalized will normalize the movement vector.
+    * This means that the character will move the same speed 
+    * diagonally as the do straight. Because pythagoras.
+    *
+    */
+    /*
+    void FixedUpdate()
+    {
+        /* 
+        * Jumping is handled by applying a constant velocity to the player for
+        * a duration of time. This is to prevent the player from snapping to
+        * their max jump height, and makes it a bit more smooth. 
+        *
+        * I THINK I have it so jump height represents the maximum number of Unity 
+        * units the player will travel upwards, but don't quote this on that.
+        */
+
+
+        /*
+        
+
+        //if the player is jumping (duh)
+        if (jump)
+        {
+            jump = false;
+            playerVerticalVelocity.y = jumpHeight;
+        }
+        else if (!controller.isGrounded)
+        {
+            //In this else if is when the player is NOT jumping but is still in the air
+            if (hardFalling)
+                playerVerticalVelocity.y += (gravity * floatingJumpModifier) * Time.deltaTime;
+            else if (Input.GetButton("Jump"))
+                playerVerticalVelocity.y += gravity * Time.deltaTime;    
+        }
+        else 
+        {
+            //This else is when the player is not jumping and is grounded.
+            playerVerticalVelocity.y = gravity * Time.deltaTime;
+        }
+        Debug.Log(playerVerticalVelocity.y);
+
+        //move = move.normalized;
+        //controller.Move(move * moveSpeed * Time.deltaTime);
+        player.Move(move);
+
+
+        // Add gravity to player's vertical velocity. 
+        //playerVerticalVelocity.y -= gravity;
+        controller.Move(playerVerticalVelocity * Time.deltaTime); 
+    }
+    */
 }
